refactor: modularize modes and actions #15
