+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/henriquehbr/svelte-typewriter/compare/v2.1.16...HEAD)
 
-- refactor: implement new architecture PoC [`9145735`](https://github.com/henriquehbr/svelte-typewriter/commit/91457350592ec0b33d5f553b0ce0566337a2c700)
+- refactor: modularize modes and actions [`ed20270`](https://github.com/henriquehbr/svelte-typewriter/commit/ed2027015f459243dbf88a282da3db178e771e91)
+- refactor: implement new architecture PoC [`c5044a2`](https://github.com/henriquehbr/svelte-typewriter/commit/c5044a207f69164aa9b18afd1680806c38bfac1b)
 
 #### [v2.1.16](https://github.com/henriquehbr/svelte-typewriter/compare/v2.1.15...v2.1.16)
