+++ example/index.html
@@ -5,9 +5,9 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
 		<title>svelte-typewriter</title>
-		<link rel="stylesheet" href="build/bundle.css" />
+		<link rel="stylesheet" href="/build/bundle.css" />
 	</head>
 	<body>
-		<script type="module" defer src="build/index.js"></script>
+		<script type="module" defer src="/build/index.js"></script>
 	</body>
 </html>
