+++ example/index.js
@@ -1,3 +1,14 @@
 import App from './App.svelte'
 
-new App({ target: document.body })
+const app = new App({
+	target: document.body
+})
+
+export default app
+
+if (import.meta.hot) {
+	import.meta.hot.dispose(() => {
+		app.$destroy()
+	})
+	import.meta.hot.accept()
+}
