+++ example/package.json
@@ -4,16 +4,19 @@
 	"main": "index.js",
 	"license": "MIT",
 	"scripts": {
-		"start": "rollup -cw"
+		"dev": "rollup -c -w",
+		"start": "sirv ./"
 	},
 	"devDependencies": {
 		"@rollup/plugin-commonjs": "^14.0.0",
 		"@rollup/plugin-node-resolve": "^8.4.0",
 		"rollup": "^2.21.0",
-		"rollup-plugin-livereload": "^1.3.0",
-		"rollup-plugin-serve": "^1.0.1",
-		"rollup-plugin-svelte": "^5.2.3",
+		"rollup-plugin-hot": "^0.1.0",
+		"rollup-plugin-svelte-hot": "^0.10.0",
 		"rollup-plugin-terser": "^6.1.0",
 		"svelte": "^3.24.0"
+	},
+	"dependencies": {
+		"sirv-cli": "^1.0.6"
 	}
 }
