+++ src/CursorStyle.svelte
